+++ web/src/containers/User.js
@@ -85,17 +85,23 @@ class User extends React.Component {
   }
 
   handleDelete() {
-    const { dispatch, user } = this.props;
-    const { selectedRowKeys, pagination } = this.state;
+    Modal.confirm({
+      title: 'Are you sure to delete ?',
+      onOk: () => {
+        const { dispatch, user } = this.props;
+        const { selectedRowKeys, pagination } = this.state;
 
-    if (selectedRowKeys.length > 0) {
-      dispatch(UserDelete(
-        map(selectedRowKeys, (i) => user.list[i].id),
-        pagination.pageSize,
-        pagination.current
-      ));
-      this.setState({ selectedRowKeys: [] });
-    }
+        if (selectedRowKeys.length > 0) {
+          dispatch(UserDelete(
+            map(selectedRowKeys, (i) => user.list[i].id),
+            pagination.pageSize,
+            pagination.current
+          ));
+          this.setState({ selectedRowKeys: [] });
+        }
+      },
+      iconType: 'exclamation-circle',
+    });
   }
 
   handleInfoShow(info) {
